Move function from rocketchat:lib to rocketchat:utils to use it in rocketchat:models
