+++ packages/rocketchat-lib/lib/getDefaultSubscriptionPref.js
@@ -1,31 +1,3 @@
-RocketChat.getDefaultSubscriptionPref = function _getDefaultSubscriptionPref(userPref) {
-	const subscription = {};
+import { getDefaultSubscriptionPref } from 'meteor/rocketchat:utils';
 
-	const {
-		desktopNotifications,
-		mobileNotifications,
-		emailNotificationMode,
-		highlights,
-	} = (userPref.settings && userPref.settings.preferences) || {};
-
-	if (Array.isArray(highlights) && highlights.length) {
-		subscription.userHighlights = highlights;
-	}
-
-	if (desktopNotifications && desktopNotifications !== 'default') {
-		subscription.desktopNotifications = desktopNotifications;
-		subscription.desktopPrefOrigin = 'user';
-	}
-
-	if (mobileNotifications && mobileNotifications !== 'default') {
-		subscription.mobilePushNotifications = mobileNotifications;
-		subscription.mobilePrefOrigin = 'user';
-	}
-
-	if (emailNotificationMode && emailNotificationMode !== 'default') {
-		subscription.emailNotifications = emailNotificationMode;
-		subscription.emailPrefOrigin = 'user';
-	}
-
-	return subscription;
-};
+RocketChat.getDefaultSubscriptionPref = getDefaultSubscriptionPref;
