+++ packages/rocketchat-utils/client/index.js
@@ -1,9 +1,11 @@
 import { t, isRtl } from '../lib/tapi18n';
 import { isChrome, isFirefox } from './lib/browsers';
+import { getDefaultSubscriptionPref } from '../lib/getDefaultSubscriptionPref';
 
 export {
 	t,
 	isRtl,
 	isChrome,
 	isFirefox,
+	getDefaultSubscriptionPref,
 };
