+++ packages/rocketchat-utils/server/index.js
@@ -1,6 +1,8 @@
 import { t, isRtl } from '../lib/tapi18n';
+import { getDefaultSubscriptionPref } from '../lib/getDefaultSubscriptionPref';
 
 export {
 	t,
 	isRtl,
+	getDefaultSubscriptionPref,
 };
